+++ Sources/Core/Renderer/GUI/GUIElementUniforms.swift
@@ -3,5 +3,5 @@ import FirebladeMath
 /// The uniforms for a GUI element.
 struct GUIElementUniforms {
   /// The element's position.
-  var position: Vec2i
+  var position: Vec2f
 }
